moved some ble functions to a new module
